requirements.txt
langchain==0.2.2
langchain_openai==0.1.8
python-dotenv==1.0.1
SQLAlchemy==2.0.30
streamlit==1.35.0
